LayoutTests/fast/dom/Element/insertAdjacentText-parameters.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description("Test that parameters to insertAdjacentText() / insertAdjacentHTML() are mandatory.");

shouldThrow("document.body.insertAdjacentText()", "'TypeError: Not enough arguments'");
shouldThrow("document.body.insertAdjacentText('beforebegin')", "'TypeError: Not enough arguments'");
shouldThrow("document.body.insertAdjacentHTML()", "'TypeError: Not enough arguments'");
shouldThrow("document.body.insertAdjacentHTML('beforebegin')", "'TypeError: Not enough arguments'");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
