Safari-609.1.11.1/LayoutTests/fast/dom/Element/insertAdjacentText-parameters-expected.txt

Test that parameters to insertAdjacentText() / insertAdjacentHTML() are mandatory.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.body.insertAdjacentText() threw exception TypeError: Not enough arguments.
PASS document.body.insertAdjacentText('beforebegin') threw exception TypeError: Not enough arguments.
PASS document.body.insertAdjacentHTML() threw exception TypeError: Not enough arguments.
PASS document.body.insertAdjacentHTML('beforebegin') threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
